Refactoring adapters.py to adapter.py to resolve name conflict
